# perf: remove to vec to sort the slice