perf: remove to vec to sort the slice commit a8157e501ae2fa82bafba76ecc44bad4a62f422d
```
part1                   time:   [123.79 ns 124.55 ns 125.43 ns]
                        change: [-20.742% -20.374% -19.969%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  2 (2.00%) high mild
  8 (8.00%) high severe

part2                   time:   [174.64 ns 175.45 ns 176.57 ns]
                        change: [+2.2454% +3.3422% +4.7322%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```